JUnit v5 refactor Util Methods into VfsTestUtils
