+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/hdfs/HdfsFileProviderTestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.hdfs;
 
+import static org.apache.commons.vfs2.VfsTestUtils.getTestDirectoryFile;
+
 import java.io.File;
 import java.io.IOException;
