+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/jar/NestedJarTestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.jar;
 
+import static org.apache.commons.vfs2.VfsTestUtils.getTestResource;
+
 import java.io.File;
 
 import org.apache.commons.vfs2.AbstractProviderTestConfig;
