+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/local/LocalProviderTestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.local;
 
+import static org.apache.commons.vfs2.VfsTestUtils.getTestDirectoryFile;
+
 import org.apache.commons.lang3.SystemUtils;
 import org.apache.commons.vfs2.AbstractProviderTestConfig;
 import org.apache.commons.vfs2.FileObject;
