+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/sftp/SftpPermissionExceptionTestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.sftp;
 
+import static org.apache.commons.vfs2.VfsTestUtils.getTestDirectory;
+
 import java.io.File;
 import java.nio.file.Paths;
