+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/tar/NestedTbz2TestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.tar;
 
+import static org.apache.commons.vfs2.VfsTestUtils.getTestResource;
+
 import org.apache.commons.vfs2.AbstractProviderTestConfig;
 import org.apache.commons.vfs2.FileObject;
 import org.apache.commons.vfs2.FileSystemManager;
