+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/temp/TemporaryProviderTestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.temp;
 
+import static org.apache.commons.vfs2.VfsTestUtils.getTestDirectoryFile;
+
 import java.io.File;
 
 import org.apache.commons.vfs2.AbstractProviderTestConfig;
