+++ commons-vfs2/src/test/java/org/apache/commons/vfs2/provider/test/GenericFileNameTestCase.java
@@ -16,6 +16,8 @@
  */
 package org.apache.commons.vfs2.provider.test;
 
+import static org.apache.commons.vfs2.VfsTestUtils.assertSameMessage;
+
 import org.apache.commons.vfs2.AbstractVfsTestCase;
 import org.apache.commons.vfs2.FileSystemException;
 import org.apache.commons.vfs2.provider.GenericFileName;
